% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelist_helpers.R
\name{add_def}
\alias{add_def}
\title{Add Definition to Codelist}
\usage{
add_def(x, template, name, ...)
}
\arguments{
\item{x}{list. Codelist definition is added to.}

\item{template}{list. Codelist template used to set default values.}

\item{name}{string. Name of definition.}

\item{...}{Add/update definition elements.}
}
\value{
list
}
\description{
Add a definition to a codelist based on a given definition template.
}
\examples{
codelist <- list() |>
  add_def(def_template(), "def")
}
