% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_to_regex.R
\name{chr_to_regex}
\alias{chr_to_regex}
\title{Make regular expression from character vector}
\usage{
chr_to_regex(x, regex_type = c("starts_with", "exact", "ends_with"))
}
\arguments{
\item{x}{character vector.}

\item{regex_type}{Type of regex to make from the input vector.}
}
\value{
string.
}
\description{
Makes a regular expression searching for strings given in the input vector.
By default, the regular expression search for strings beginning with the
strings from the input vector, but it can also by specified to look for exact
matches or strings ending with the strings from the input vector.
}
\examples{
chr_to_regex(c("abc", "def"))
}
