% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelist_helpers.R
\name{codelist_to_df}
\alias{codelist_to_df}
\title{Convert Codelist to data.frame}
\usage{
codelist_to_df(x)
}
\arguments{
\item{x}{list. Codelist.}
}
\value{
data.frame
}
\description{
Converts a codelist into a data.frame
}
\examples{
 template <- def_template()
 template["def_type"] <- "yes_no"
 template["def_period"] <- "before"
 codelist <- list() |>
   add_def(template, "def1") |>
   add_def(template, "def2") |>
   update_codes("def1", "icd8", c("123", "124")) |>
   update_codes("def2", "icd10", c("A", "B", "C"), include = FALSE) |>
   codelist_to_df()
}
