% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr_extract.R
\name{cpr_extract}
\alias{cpr_extract}
\title{Extract birthdate or biological sex encoded in CPR-numbers}
\usage{
cpr_extract(x, extract = c("birthdate", "sex"), test_modulus11 = FALSE)
}
\arguments{
\item{x}{Character vector. Vector of CPR-numbers}

\item{extract}{String. "birthdate" or "sex" depending on what is to be
extracted. Sex will be returned as a zero/one vector, where one's indicates
males.}

\item{test_modulus11}{logical. Test if CPR-numbers fulfill the modulus 11 test?
If true, will return error if the modulus 11 test fails for one or more
CPR-numbers}
}
\value{
vector
}
\description{
Extract information about dates of birth or biological sex, from Danish
CPR-numbers.
}
\details{
More information on Danish CPR-numbers (in Danish) can be found at
https://cpr.dk/cpr-systemet/opbygning-af-cpr-nummeret.
}
\examples{
cpr_extract("0101001234")
}
