% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelist_helpers.R
\name{def_template}
\alias{def_template}
\title{Make a Definition Template}
\usage{
def_template(
  code_types = list(icd8 = "", icd10 = "", atc = "", bio_npu = "", bio_ana = ""),
  ...
)
}
\arguments{
\item{code_types}{list.}

\item{...}{Update/add definition elements.}
}
\value{
list
}
\description{
Make a definition template. Intended to be used together with the other
codelist helper functions to facilitate creating a codelist.
}
\examples{
def_template()
}
