% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variable_helpers.R
\name{define_all_variables}
\alias{define_all_variables}
\title{Define All Variables}
\usage{
define_all_variables(
  studypop,
  var_dat,
  codelist,
  id_var = "id",
  index_date_var = "index_date",
  date_var = "date",
  code_var = "code",
  code_type_var = "code_type",
  result_var = "result"
)
}
\arguments{
\item{studypop}{data.frame. Study population.}

\item{var_dat}{data.frame. Data used to define variables.}

\item{codelist}{list. Codelist.}

\item{id_var}{string. Name of id variable in \code{studypop} and \code{var_dat}.}

\item{index_date_var}{string. Name of index date variable in \code{studypop}.}

\item{date_var}{string. Name of variable in \code{var_dat} with dates
corresponding \code{code_var}.}

\item{code_var}{string. Name of code variable in in \code{var_dat}.}

\item{code_type_var}{string. Name of code type variable in \code{var_dat}.}

\item{result_var}{string. Name of result variable in in \code{var_dat}.}
}
\value{
data.table
}
\description{
Define all variables in a codelist using a study population and
variable data.
}
\examples{
\dontrun{
TODO: add example
}
}
