% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variable_helpers.R
\name{define_variable}
\alias{define_variable}
\title{Define Variable}
\usage{
define_variable(
  x,
  def,
  id_var = "id",
  index_date_var = "index_date",
  date_var = "date",
  code_var = "code",
  code_type_var = "code_type",
  result_var = "result"
)
}
\arguments{
\item{x}{data.frame. Variable data.}

\item{def}{list. Codelist definition.}

\item{id_var}{string. Name of id variable in \code{x}.}

\item{index_date_var}{string. Name of index date variable in \code{x}.}

\item{date_var}{string. Name of variable in \code{x} with dates
corresponding \code{code_var}.}

\item{code_var}{string. Name of code variable in \code{x}.}

\item{code_type_var}{string. Name of code type variable in \code{x}.}

\item{result_var}{string. Name of result variable in \code{x}.}
}
\value{
data.table
}
\description{
Defines a variable based on variable data and a variable definition.
}
\examples{
\dontrun{
TODO: add example
}
}
