% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_codes.R
\name{.clean_codes}
\alias{.clean_codes}
\title{Clean codes}
\usage{
.clean_codes(
  x,
  remove_periods = TRUE,
  check_if_clean = TRUE,
  range_symbols = "[-:]",
  sep_symbols = "[,\\\\|;\\\\s]"
)
}
\arguments{
\item{x}{character. Vector of codes and code ranges to tidy.}

\item{remove_periods}{logical. Remove periods, assumed to be part of codes?}

\item{check_if_clean}{logical. Check if return vector has been properly
cleaned, ie contains no non-alphanumeric symbols
(excluding dashes and spaces)?}

\item{range_symbols}{string. \code{grepl} pattern used to identify range symbols}

\item{sep_symbols}{string. \code{grepl} pattern used to identify code separator
symbols.}
}
\value{
character
}
\description{
Cleans a vector of codes and/or code ranges by streamlining symbols used for
code ranges (dashes) and splitting the vector so that each code(range) is in
a separate element.
}
\details{
By default the function removes any period symbols present in the input,
assumed to be part of the codes. The function will also throw an error if
the return vector contains unexpected non-alphanumeric values, assumed to
mean that the input character vector includes symbols that have not been
properly handled.

Note that the function is case-insensitive and will convert any lowercase
letters to uppercase.
}
\examples{
\dontrun{.clean_codes(c("A.1", "123", "A123,B43; K45 K67 A1-B2", "B:D"))}
}
\keyword{internal}
