% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variable_helpers.R
\name{.code_filter_exp}
\alias{.code_filter_exp}
\title{Make Code Filter String Expression}
\usage{
.code_filter_exp(def, code_type_var = "code_type", code_var = "code")
}
\arguments{
\item{def}{list. codelist definition.}

\item{code_type_var}{string. Name of code type variable.}

\item{code_var}{string. Name of code variable}
}
\value{
string
}
\description{
Takes a codelist definition and creates a string that can be used as an
expression to restrict variable data to rows with codes relevant for
the definition.
}
\examples{
\dontrun{
 def <- list() |>
 add_def(def_template(), "var") |>
   update_codes("var", "icd8", c("1", "2"))
 def <- def$var
 .code_filter_exp(def)
}
}
\keyword{internal}
