% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_codes.R
\name{.collapse_codes}
\alias{.collapse_codes}
\title{Collapse codes}
\usage{
.collapse_codes(x, sort_codes = TRUE)
}
\arguments{
\item{x}{character.}

\item{sort_codes}{boolean. Sort vector of collapsed codes?}
}
\value{
character
}
\description{
Collapse vector of codes, if possible.
}
\details{
Attempts to collapse a vector of codes into a smaller set of codes that is
equivalent. For example, if the input is the codes:
A0, A012, A1, A2, A3, A4, A5, A6, A7, A8, A9, B, B1
then the function would detect that A012 is a subcode of A0 and B1 of B, so
they can be removed. Furthermore
A0, A1, A2, A3, A4, A5, A6, A7, A8, A9
is the set of all subcode chapters in A, so we can remove all these codes
and simply include A instead. The output of the function in this example
would simply be the codes: A, B.

Note that the function is case-insensitive and will return character codes
in uppercase.
}
\examples{
\dontrun{.collapse_codes(c(paste0("A", 0:9), "B", "B1"))}
}
\keyword{internal}
