% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_codes.R
\name{.expand_code_range}
\alias{.expand_code_range}
\title{Expand code range}
\usage{
.expand_code_range(x, range_sep = "-")
}
\arguments{
\item{x}{Character. Vector of length 1. Code range to be expanded}

\item{range_sep}{string. Range separator symbol. Must be length 1.}
}
\value{
character
}
\description{
Expand a code range into a a character vector with the codes the range
represents.
}
\examples{
\dontrun{.expand_code_range("A-C")}
}
\keyword{internal}
