% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variable_helpers.R
\name{.filter_var_def_dat}
\alias{.filter_var_def_dat}
\title{Filter Variable Data}
\usage{
.filter_var_def_dat(
  var_dat,
  def,
  index_date = "index_date",
  date_var = "date",
  code_var = "code",
  code_type_var = "code_type"
)
}
\arguments{
\item{var_dat}{data.frame. Date on variable}

\item{def}{list. Codelist definition.}

\item{index_date}{string. Name of index date variable in \code{var_dat}.}

\item{date_var}{string. Name of variable in \code{var_dat} with dates
corresponding \code{code_var}.}

\item{code_var}{string. Name of code variable in \code{var_dat}.}

\item{code_type_var}{string. Name of code type variable \code{var_dat}.}
}
\value{
data.table
}
\description{
Takes a variable definition and data on the variable and restricts the
data to only include rows that are relevant for defining the variable.
}
\examples{
\dontrun{TODO: add example}
}
\keyword{internal}
