% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_codes.R
\name{.next_code}
\alias{.next_code}
\title{Find next code}
\usage{
.next_code(x)
}
\arguments{
\item{x}{character}
}
\value{
character
}
\description{
Find the "next" code.
}
\examples{
\dontrun{.next_code("AB2")}
}
\keyword{internal}
