% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variable_helpers.R
\name{.period_filter_exp}
\alias{.period_filter_exp}
\title{Make Period Filter String Expression}
\usage{
.period_filter_exp(def, index_date = "index_date", date_var = "date")
}
\arguments{
\item{def}{list. codelist definition.}

\item{index_date}{string. Name of index date variable.}

\item{date_var}{string. Name of variable with code dates.}
}
\value{
string
}
\description{
Takes a codelist definition and creates a string that can be used as an
expression restrict variable data to rows with codes relevant for the
definition.
}
\examples{
\dontrun{
TODO: add example
}
}
\keyword{"before")}
\keyword{"def_length",}
\keyword{"def_period",}
\keyword{"var")}
\keyword{.period_filter_exp(def)}
\keyword{10)}
\keyword{<-}
\keyword{add_def(def_template(),}
\keyword{def}
\keyword{def$var}
\keyword{internal}
\keyword{list()}
\keyword{update_def("var",}
\keyword{|>}
