% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_codes.R
\name{.validate_code_range}
\alias{.validate_code_range}
\title{Validate code range}
\usage{
.validate_code_range(x, range_sep = "-")
}
\arguments{
\item{x}{character. Vector of length 1}

\item{range_sep}{string. Range separator symbol. Must be length 1.}
}
\value{
character
}
\description{
Validates a code range. Returns the code range, or throws an error if the
input is not a proper code range.
}
\details{
A proper code range is on the form "from-to", where from and to is the same
length, and from is "lower than or equal to" to, eg "A0-A4", "1Z-2B",
"ABC-DAG" are valid ranges, but "1-10" and "B1-A9" are not.
}
\examples{
\dontrun{.validate_code_range("A-C")}
}
\keyword{internal}
