% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_date.R
\name{is_date}
\alias{is_date}
\title{Check if object is a Date object}
\usage{
is_date(x)
}
\arguments{
\item{x}{object}
}
\value{
logical
}
\description{
Check if object is a Date object
}
\examples{
is_date(as.Date("2000-01-01"))
}
