% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_integer.R
\name{is_integer}
\alias{is_integer}
\title{Check if numeric}
\usage{
is_integer(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{a numeric vector}

\item{tol}{tolerance}
}
\value{
logical vector
}
\description{
Check if each element of a numeric vector can be interpreted as integer
numbers.
}
\examples{
is_integer(1)
is_integer(1.1)
}
