% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_variable_name.R
\name{is_valid_variable_name}
\alias{is_valid_variable_name}
\title{Check if valid variable name}
\usage{
is_valid_variable_name(x, allow_reserved = FALSE, unique = FALSE)
}
\arguments{
\item{x}{character vector}

\item{allow_reserved}{boolean}

\item{unique}{Check if variable name in \code{x} is unique}
}
\value{
logical vector
}
\description{
Check if vector of candidate variable names are valid variable names.
}
\examples{
is_valid_variable_name("valid_variable_name")
is_valid_variable_name("_not_valid")
}
