% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_tbl.R
\name{mask_tbl}
\alias{mask_tbl}
\title{Mask table}
\usage{
mask_tbl(x)
}
\arguments{
\item{x}{tbl_summary object.}
}
\value{
tbl_summary object
}
\description{
Mask person-sensitive data in a table
}
\details{
Masks person-sensitive data in a tbl_summary object made using the gtsummary
package. Function should work as intended, but is very experimental and
insufficiently undocumented and tested. The whole approach of editing
the table_body part of the tbl_summary object is probably also fundamentally
flawed approach. This is pretty hack-ish.
}
\examples{
\dontrun{
TODO: add example
}
}
