% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_periods.R
\name{smooth_periods}
\alias{smooth_periods}
\title{Smooth periods}
\usage{
smooth_periods(x, start = "start", end = "end", group = NULL, max_gap = 1L)
}
\arguments{
\item{x}{data.table}

\item{start}{string. Variable name of numeric/date variable with start dates.}

\item{end}{string. Variable names of numeric/date variable with end dates.}

\item{group}{character vector. Names of grouping variables.}

\item{max_gap}{Non-negative integer. Maximum allowed gap between time-periods
when smoothing.}
}
\value{
data.table
}
\description{
Smooth data with time periods, so that data lines with periods in
continuation of each others are combined into a single data line.
}
\examples{
2 + 2
}
