% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_codes.R
\name{tidy_codes}
\alias{tidy_codes}
\title{Tidy codes}
\usage{
tidy_codes(x, range_symbols = "[:-]", code_sep_symbols = "[,\\\\|;\\\\s]")
}
\arguments{
\item{x}{character. Vector of codes and code ranges.}

\item{range_symbols}{string. \code{grepl} pattern used to identify range symbols.}

\item{code_sep_symbols}{string. \code{grepl} pattern used to identify code
separator symbols.}
}
\value{
character
}
\description{
Tidy a character vector of codes and code ranges by cleaning the input,
resolving code ranges, collapsing the codes into an equivalent minimal
set of codes that is sorted, and where there is only one code in each
element of the vector.
}
\details{
It is a time-consuming and error-prone process to manually make
code definitions based on a list of codes in a protocol, eg ICD-10 codes
used to define diseases in registry data. This utility function can help
facilitate the process. All codes in a codebook definition can simply be
copy+pasted into a string, and the function will clean everything up.

The function only accepts code ranges where the start and end of the range
is specified at the same level, eg "AB-AD" is ok, but "20-400" is not. This
is because this code ranges is ambiguous: sometimes this actually means
200-400.
}
\examples{
# Simple neat input
tidy_codes("A12-A19, B10")
# Messy complex input
tidy_codes(c("A2;B12-B14 123-125", "G98-H01", "C09-C41"))
}
