% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelist_helpers.R
\name{update_codes}
\alias{update_codes}
\title{Update Codes}
\usage{
update_codes(x, name, type, codes, include = TRUE)
}
\arguments{
\item{x}{list. Codelist in which a definition is to be updated.}

\item{name}{string. Name of definition to update.}

\item{type}{string. Type of codes to update.}

\item{codes}{Character. Vector of codes.}

\item{include}{Boolean. Are the codes to update the inclusion codes (TRUE) or
the exclusion codes (FALSE)}
}
\value{
list
}
\description{
Update a specific type of codes used in a codelist definition.
}
\examples{
codelist <- list() |>
  add_def(def_template(), "def") |>
  update_codes("def", "icd8", c("123", "124"))
}
