% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelist_helpers.R
\name{update_def}
\alias{update_def}
\title{Update Codelist Definition}
\usage{
update_def(x, name, var, value)
}
\arguments{
\item{x}{list. Codelist in which a definition is to be updated}

\item{name}{string. Name of definition to update.}

\item{var}{string. Definition element to update.}

\item{value}{Value assigned to \code{var}.}
}
\value{
list
}
\description{
Update a definition in a codelist.
}
\examples{
codelist <- list() |>
  add_def(def_template(), "def") |>
  update_def("def", "manual_def", TRUE)
}
